Collect timezone and allow userAgentData to be overridden (#956)
