+++ .changeset/cool-peaches-pay.md
@@ -1,0 +1,5 @@
+---
+'@segment/analytics-next': minor
+---
+
+Set timezone and allow userAgentData to be overridden
